Install zsh completion with bazelisk
This isn't ideal, as we should really be dynamically updating the
completion based on the bazel version in the current repo, but that
would take a lot more changes to bazelisk to provide a stable directory
for that.

This at least gives reasonable completions for those who keep to the
latest release.